Distinguish “target” and “current” views.
The target view represents the programmatically-set view; the current view
represents the active element’s view as displayed, as determined by the last
application of the zoom behavior and subsequently modified through interaction.

The target view is NOT updated through interaction. This is because the zoom
behavior may be applied to multiple elements, and the current view of each
element may differ; it is only set programmatically.

Related #3.