Remove zoom.center.
The problem is that it’s too vague or specific a concept. Does it apply just to
wheeling, or also to double clicks? Does it affect zoom transitions and clamping
by the scale extent?

Originally this feature was conceived of letting you zoom in and out, but not
translate. But it doesn’t work well for this, because it doesn’t work for touch,
and because you still have to unregister the dblclick and mousedown listeners. I
need to design a more general replacement for this, such as a hook for
constraining the zoom transform arbitrarily.