Use touches rather than changedTouches in touchstarted
It seems that the behaviour of Chrome on Android differs from the
behaviour of Safari on iOS in the following respect: when a multitouch
event is initiated by touching two or more fingers to the screen
near-simultaneously, iOS Safari will fire a single touchstart event
that includes all the touches (as both event.touches and event.changedTouches)
whereas Android Chrome will fire a sequence of touchstart events:

- one with touches.length == 1 and changedTouches.length == 1
- followed by one with touches.length == 2 and changedTouches.length == 1

In other words, the behaviour on Android Chrome when the screen is
touched with two fingers simultaneously is the same as the behaviour
on iOS Safari when you touch it first with one finger and then a second.

The change in PR #169 makes it possible to ignore single-touch gestures
on iOS Safari, but this fails on Android Chrome because the first touch
is never recorded. When the second touch is processed, the changedTouches
array contains only the second touch.

This commit changes the behaviour of the touchstarted handler to consider
all the touches in the event, rather than just the new touches.